requirements.txt
pytorch
torchvision
hydra-core
meerkat-ml==0.2.0
imageio
dosma
pytorch-lightning
wandb
segmentation-models-pytorch
albumentations